# public/docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "OMPay API",
        "description": "Documentation de l'API OMPay (Orange Money Pay)",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Serveur local"
        },
        {
            "url": "https://ompay.onrender.com/api/v1",
            "description": "Serveur de production"
        }
    ],
    "paths": {
...
        "/distributeur/depot": {
            "post": {
                "tags": [
                    "Distributeur"
                ],
                "summary": "Effectuer un dépôt sur le compte d'un client (montant positif)",
                "operationId": "depotDistributeur",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "numero",
                                    "montant"
                                ],
                                "properties": {
                                    "numero": {
                                        "type": "string",
                                        "example": "+221771428150"
                                    },
                                    "montant": {
                                        "type": "number",
                                        "example": 3000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Dépôt effectué",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "type": {
                                                    "type": "string",
                                                    "example": "depot"
                                                },
                                                "montant": {
                                                    "description": "Montant positif",
                                                    "type": "number",
                                                    "example": 3000
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "terminé"
                                                },
                                                "date_transaction": {
                                                    "type": "string",
                                                    "example": "2025-11-16 12:00:00"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Dépôt effectué"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/distributeur/retrait": {
            "post": {
                "tags": [
                    "Distributeur"
                ],
                "summary": "Effectuer un retrait sur le compte d'un client (montant négatif)",
                "operationId": "retraitDistributeur",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "qr_code",
                                    "montant"
                                ],
                                "properties": {
                                    "qr_code": {
                                        "type": "string",
                                        "example": "QRCODE5442"
                                    },
                                    "montant": {
                                        "type": "number",
                                        "example": 3000
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Retrait effectué",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "type": {
                                                    "type": "string",
                                                    "example": "retrait"
                                                },
                                                "solde_retiré": {
                                                    "description": "Montant retiré (positif dans la réponse, négatif en base)",
                                                    "type": "number",
                                                    "example": 3000
                                                },
                                                "solde_restant": {
                                                    "type": "number",
                                                    "example": 12000
                                                },
                                                "numero_distributeur": {
                                                    "type": "string",
                                                    "example": "+221770000001"
                                                },
                                                "frais": {
                                                    "type": "number",
                                                    "example": 0
                                                },
                                                "telephone_client": {
                                                    "type": "string",
                                                    "example": "+221771428150"
                                                },
                                                "status": {
                                                    "type": "string",
                                                    "example": "terminé"
                                                },
                                                "date_transaction": {
                                                    "type": "string",
                                                    "example": "2025-11-16 12:00:00"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Retrait effectué"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
...
}